Refactoring postponed till Canada. Everything is working as it should. No app breaking bugs. Will be sent out to be tested.

+++ lib/utilityClasses.dart
@@ -10,6 +10,6 @@ class GlobalSettings
 {  //GlobalSettings(this.gridLayout, this.numberOfHeadlines, this.readSpeed);
 
   bool gridLayout;
-  int numberOfHeadlines;
+  double numberOfHeadlines;
   double readSpeed;
 }
